refactor(settings): move run-health accessors, consts, validator to settings_health.go (PRD #1021 M3)
Pure same-package motion: HealthEnabled and the five integer-seconds accessors,
the healthSecondsMin/Max bounds const block, and validateHealthSeconds move
verbatim from settings.go into settings_health.go. CapabilityAwareScheduling
stays in settings.go (domainless one-off). Exported surface byte-identical via
go doc -all; Validate switch cases stay and call validateHealthSeconds cross-file.
